smart_waste_ui/requirements.txt
asgiref==3.8.1
Django==5.2.2
kafka-python==2.2.10
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
djangorestframework

